# artifacts/api-server/openapi.yaml
openapi: 3.0.3
info:
  title: ISM Finance API
  version: 1.2.0
  description: API REST do SaaS financeiro com painel JWT e integração Power BI (v1 token).
servers:
  - url: http://localhost:3000/api
    description: Local
...
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      operationId: postAuthLogin
      summary: Login com email e senha
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, senha]
              properties:
                email: { type: string, format: email }
                senha: { type: string }
      responses:
        "200":
          description: Login efetuado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelopeAuthLogin"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /auth/me:
    get:
      tags: [Auth]
      operationId: getAuthMe
      summary: Usuário autenticado + permissões
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Usuário atual
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelopeMe"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    bearerV1Token:
      type: http
      scheme: bearer
      description: Token estático validado na tabela `tokens_api`.
  responses:
    UnauthorizedError:
      description: Não autenticado
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiEnvelopeError"
...
  schemas:
    ApiErrorItem:
      type: object
      properties:
        code: { type: string }
        message: { type: string }
        details: {}
    ApiEnvelopeGeneric:
      type: object
      properties:
        data: {}
        meta:
          nullable: true
          type: object
          additionalProperties: true
        errors:
          nullable: true
          type: array
          items:
            $ref: "#/components/schemas/ApiErrorItem"
...
    ApiEnvelopeAuthLogin:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelopeGeneric"
        - type: object
          properties:
            data:
              type: object
              properties:
                token: { type: string }
                user:
                  $ref: "#/components/schemas/Usuario"
            meta:
              type: object
              properties:
                tokenType: { type: string, example: Bearer }
                expiresIn: { type: string, example: 24h }
    ApiEnvelopeMe:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelopeGeneric"
        - type: object
          properties:
            data:
              type: object
              properties:
                user:
                  $ref: "#/components/schemas/Usuario"
                permissoes:
                  type: array
                  items: { type: string }
    ApiEnvelopeError:
      type: object
      properties:
        data:
          nullable: true
        meta:
          nullable: true
        errors:
          type: array
          items:
            $ref: "#/components/schemas/ApiErrorItem"
    Usuario:
      type: object
      properties:
        id: { type: integer }
        nome: { type: string }
        email: { type: string, format: email }
...